requirements.txt
faiss-cpu
jsonlines
mistralai
numpy
PyYAML
sentence_transformers
matplotlib
networkx